requirements/requirements.3.11.txt
anyio==4.3.0
appdirs==1.4.4
certifi>=2024.2.2
ecdsa==0.18.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
pyasn1==0.5.1
python-dateutil==2.8.2
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
typeguard==4.1.5
typing_extensions==4.10.0
